+++ modules/stitching/src/blenders.cpp
@@ -267,7 +267,7 @@ static bool ocl_MultiBandBlender_feed(InputArray _src, InputArray _weight,
            ocl::KernelArg::ReadWrite(_dst_weight.getUMat())
            );
 
-    size_t globalsize[2] = {src.cols, src.rows };
+    size_t globalsize[2] = {(size_t)src.cols, (size_t)src.rows };
     return k.run(2, globalsize, NULL, false);
 }
 #endif
@@ -469,7 +469,7 @@ static bool ocl_normalizeUsingWeightMap(InputArray _weight, InputOutputArray _ma
            ocl::KernelArg::ReadOnly(_weight.getUMat())
            );
 
-    size_t globalsize[2] = {mat.cols, mat.rows };
+    size_t globalsize[2] = {(size_t)mat.cols, (size_t)mat.rows };
     return k.run(2, globalsize, NULL, false);
 }
 #endif
